- change days cache in ScheduleDao from LiveData to a hot Flow - move ScheduleDao access to a ViewModel in TracksFragment - rename TracksViewModel to TracksListViewModel.
