+++ app/src/main/java/be/digitalia/fosdem/fragments/TracksFragment.kt
@@ -6,22 +6,22 @@ import android.view.View
 import androidx.core.content.edit
 import androidx.core.view.isVisible
 import androidx.fragment.app.Fragment
+import androidx.fragment.app.viewModels
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleEventObserver
 import androidx.recyclerview.widget.RecyclerView
 import androidx.viewpager2.adapter.FragmentStateAdapter
 import androidx.viewpager2.widget.ViewPager2
 import be.digitalia.fosdem.R
-import be.digitalia.fosdem.db.ScheduleDao
 import be.digitalia.fosdem.model.Day
 import be.digitalia.fosdem.utils.enforceSingleScrollDirection
 import be.digitalia.fosdem.utils.instantiate
 import be.digitalia.fosdem.utils.recyclerView
 import be.digitalia.fosdem.utils.viewLifecycleLazy
+import be.digitalia.fosdem.viewmodels.TracksViewModel
 import com.google.android.material.tabs.TabLayout
 import com.google.android.material.tabs.TabLayoutMediator
 import dagger.hilt.android.AndroidEntryPoint
-import javax.inject.Inject
 
 @AndroidEntryPoint
 class TracksFragment : Fragment(R.layout.fragment_tracks), RecycledViewPoolProvider {
@@ -33,8 +33,7 @@ class TracksFragment : Fragment(R.layout.fragment_tracks), RecycledViewPoolProvi
         val tabs: TabLayout = view.findViewById(R.id.tabs)
     }
 
-    @Inject
-    lateinit var scheduleDao: ScheduleDao
+    private val viewModel: TracksViewModel by viewModels()
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -52,7 +51,7 @@ class TracksFragment : Fragment(R.layout.fragment_tracks), RecycledViewPoolProvi
             requireActivity().getPreferences(Context.MODE_PRIVATE).getInt(PREF_CURRENT_PAGE, -1)
         } else -1
 
-        scheduleDao.days.observe(viewLifecycleOwner) { days ->
+        viewModel.days.observe(viewLifecycleOwner) { days ->
             holder.run {
                 daysAdapter.days = days
